put the new article component to a single component
